improve code organization, sort VisitXXX methods alphabetically
